+++ README.md
@@ -1,3 +1,3 @@
-[![CircleCI](https://circleci.com/gh/snowdrop/spring-boot-circuit-breaker-booster.svg?style=shield)](https://circleci.com/gh/snowdrop/spring-boot-circuit-breaker-booster)
+[![CircleCI](https://circleci.com/gh/snowdrop/circuit-breaker-example.svg?style=shield)](https://circleci.com/gh/snowdrop/circuit-breaker-example)
 
 https://appdev.openshift.io/docs/spring-boot-runtime.html#mission-circuit-breaker-spring-boot
