+++ greeting-service/.openshiftio/application.yaml
@@ -110,7 +110,7 @@ objects:
       app: spring-boot-circuit-breaker-greeting
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-circuit-breaker-greeting
   spec:
     ports:
@@ -121,7 +121,7 @@ objects:
     selector:
       app: spring-boot-circuit-breaker-greeting
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
@@ -129,7 +129,7 @@ objects:
       app: spring-boot-circuit-breaker-greeting
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-circuit-breaker-greeting
   spec:
     replicas: 1
@@ -137,7 +137,7 @@ objects:
     selector:
       app: spring-boot-circuit-breaker-greeting
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -148,7 +148,7 @@ objects:
           app: spring-boot-circuit-breaker-greeting
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -200,7 +200,7 @@ objects:
       app: spring-boot-circuit-breaker-greeting
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-circuit-breaker-greeting
   spec:
     port:
