+++ greeting-service/src/main/java/dev/snowdrop/example/ExampleApplication.java
@@ -14,19 +14,21 @@
  * limitations under the License.
  */
 
-package io.openshift.booster;
+package dev.snowdrop.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
 
 /**
  * Entry point to the application.
  */
 @SpringBootApplication
-public class BoosterApplication {
+@EnableCircuitBreaker
+public class ExampleApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(BoosterApplication.class, args);
+        SpringApplication.run(ExampleApplication.class, args);
     }
 
 }
