+++ pom.xml
@@ -22,12 +22,12 @@
     <artifactId>booster-parent</artifactId>
     <version>23</version>
   </parent>
-  <groupId>io.openshift.booster</groupId>
-  <artifactId>spring-boot-circuit-breaker-parent</artifactId>
+  <groupId>dev.snowdrop.example</groupId>
+  <artifactId>circuit-breaker</artifactId>
   <version>1.5.19-3-SNAPSHOT</version>
   <packaging>pom</packaging>
-  <name>Spring Boot - Circuit Breaker Booster</name>
-  <description>Spring Boot - Circuit Breaker Booster</description>
+  <name>Spring Boot - Circuit Breaker Example</name>
+  <description>Spring Boot - Circuit Breaker Example</description>
   <properties>
     <arquillian.version>1.4.0.Final</arquillian.version>
     <awaitility.version>3.1.0</awaitility.version>
